finished all

+++ APCSA2020/src/Unit2/Line.java
@@ -8,7 +8,7 @@ public class Line
 {
 	public static double getSlope( int x1, int y1, int x2, int y2 )
 	{
-		double slope = (y2-y1)/(x2-x1);
+		double slope = ((double)y2-(double)y1)/((double)x2-(double)x1);
 		return slope;
 	}
